# backend/gwisi40server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ISI 4.0 server API",
        "title": "ISI 4.0 server API",
        "contact": {
            "name": "ICTS",
            "url": "http://www.grupoicts.com.br"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/server": {
            "post": {
                "description": "Creates a server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Creates a server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add server",
                        "name": "type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelsswagger.Server"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelsswagger.Server"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Missing"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "No"
                        }
                    }
                }
            }
        },
        "/api/v1/server/{id}": {
            "get": {
                "description": "Get server by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Get server by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Server"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "No"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a server",
                "tags": [
                    "Servers"
                ],
                "summary": "Deletes a server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Server"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "Cannot"
                        }
                    }
                }
            }
        },
        "/api/v1/servers": {
            "get": {
                "description": "Get all servers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Get all servers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Server"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "No"
                        }
                    }
                }
            }
        },
        "/api/v1/type": {
            "post": {
                "description": "Creates a type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Types"
                ],
                "summary": "Creates a type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add type",
                        "name": "type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelsswagger.Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Type"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Missing"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    }
                }
            }
        },
        "/api/v1/type/{id}": {
            "get": {
                "description": "Get type by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Types"
                ],
                "summary": "Get type by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Type"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "No"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a type",
                "tags": [
                    "Types"
                ],
                "summary": "Deletes a type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Type"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "Cannot"
                        }
                    }
                }
            }
        },
        "/api/v1/types": {
            "get": {
                "description": "Get all types",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Types"
                ],
                "summary": "Get all types",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Type"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "No"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Protocol": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "config": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "md5": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "models.Server": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "protocol": {
                    "$ref": "#/definitions/models.Protocol"
                }
            }
        },
...
        "models.Type": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "modelsswagger.Server": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "protocolID": {
                    "type": "string"
                }
            }
        },
        "modelsswagger.Type": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                }
            }
        }
    }
}